metrics: time: launch times
Add script that caputes data points about the time it takes
to run a container through a create/run/delete cycle.

The script uses `date` on the host and in the container to gather
much of the info - it presumes the date in the container is
(reasonably) in sync with the host.

Signed-off-by: Graham whaley <graham.whaley@intel.com>